Move quad_tile.cpp
quad_tile.cpp is specific to apidb backends, so move it there
